creating a hook for component store stuff
